backend/requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
motor==3.3.2
pymongo==4.6.1
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
python-dotenv==1.0.0
email-validator==2.1.0
numpy==1.26.2
opencv-python-headless==4.8.1.78
Pillow==10.1.0
slowapi==0.1.9
aiosmtplib==3.0.1
google-generativeai==0.8.3

# WebSocket and Caching
websockets==12.0
redis==5.0.1

# OAuth2
authlib==1.3.0
itsdangerous==2.1.2

# Testing Dependencies
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.25.2

